+++ navigation/DonorStack/Donate/index.tsx
@@ -1,6 +1,6 @@
 import { createStackNavigator } from '@react-navigation/stack';
 import * as React from 'react';
-import { DonateReviewScreen, DonateAddressScreen, DonateSchedulePickupScreen, DonateListscreen } from '../../../screens/DonationForm';
+import { DonateReviewScreen, DonateAddressScreen, DonateReviewDonationCart, DonateListscreen } from '../../../screens/DonationForm';
 import { NewDishFormScreen, DishProfileScreen, DishSearchScreen } from '../../../screens/Dishes';
 import DonateHomePage from '../../../screens/DonateTab';
 import DishSearch from '../../../screens/Dishes/DishSearchScreen';
@@ -51,12 +51,12 @@ function DonateTab() {
       />
       <DonationScreenStack.Screen
         name="SchedulePickupScreen"
-        component={DonateSchedulePickupScreen}
+        component={DonateReviewDonationCart}
         options={{ headerTitle: 'Donate' }}
       />
       <DonationScreenStack.Screen
         name="ReviewCartScreen"
-        component={DonateReviewScreen}
+        component={DonateReviewDonationCart}
         options={{ headerTitle: 'Donate' }}
       />
       <DonationScreenStack.Screen
