+++ screens/DonationForm/index.ts
@@ -1,4 +1,4 @@
 export { default as DonateAddressScreen } from './AddressScreen';
 export { default as DonateReviewScreen } from './ReviewScreen/ReviewCart';
-export { default as DonateSchedulePickupScreen } from './ReviewScreen/ReviewCart';
+export { default as DonateReviewDonationCart } from './ReviewScreen/ReviewCart';
 export { default as DonateListscreen } from './DonationList';
